+++ CHANGES.md
@@ -27,7 +27,8 @@ v0.3
 - eliminated need for ProxiedSession (telling subject about stopped sessions)
 - removed Credential class, refactored credential collection to a dict
 - refactored SessionKey and WebSessionKey from class to namedtuple
- 
+- refactored SessionContext to a dict
+
 Yosai uses [GitHub's Releases feature](https://github.com/blog/1547-release-your-software) for its changelogs.
 
 See [the Releases section of our GitHub project](https://github.com/YosaiProject/yosai/releases) for changelogs for each release version of Yosai.